separate crypto handler from ed25519
