+++ crates/sui-cluster-test/src/main.rs
@@ -16,7 +16,7 @@ use sui_json::SuiJsonValue;
 use sui_json_rpc_api::rpc_types::{GetObjectDataResponse, SuiExecutionStatus, TransactionResponse};
 use sui_types::{
     base_types::{encode_bytes_hex, ObjectID, SuiAddress},
-    crypto::get_key_pair,
+    crypto::KeyPair,
     gas_coin::GasCoin,
     messages::{Transaction, TransactionData},
     object::Owner,
@@ -63,7 +63,7 @@ impl ClusterTest {
         assert!(!coins.is_empty(), "Not enough gas objects to run test.");
         let signer = self.context.address;
         let wallet_context = self.wallet_context();
-        let (receipent_addr, _) = get_key_pair();
+        let (receipent_addr, _) = KeyPair::get_key_pair();
         let obj_to_transfer = coins.remove(0);
         let data = wallet_context
             .gateway
