+++ crates/sui-config/src/swarm.rs
@@ -8,7 +8,7 @@ use std::num::NonZeroUsize;
 use std::path::Path;
 use std::sync::Arc;
 use sui_types::committee::Committee;
-use sui_types::crypto::{get_key_pair_from_rng, KeyPair};
+use sui_types::crypto::{KeyPair};
 
 /// This is a config that is used for testing or local use as it contains the config and keys for
 /// all validators
@@ -56,7 +56,7 @@ impl NetworkConfig {
     /// Generate a fullnode config based on this `NetworkConfig`. This is useful if you want to run
     /// a fullnode and have it connect to a network defined by this `NetworkConfig`.
     pub fn generate_fullnode_config(&self) -> NodeConfig {
-        let key_pair = Arc::new(get_key_pair_from_rng(&mut OsRng).1);
+        let key_pair = Arc::new(KeyPair::get_key_pair_from_rng(&mut OsRng).1);
         let validator_config = &self.validator_configs[0];
 
         let mut db_path = validator_config.db_path.clone();
