+++ crates/sui-core/src/checkpoints/tests/checkpoint_tests.rs
@@ -17,7 +17,6 @@ use std::{collections::HashSet, env, fs, path::PathBuf, sync::Arc, time::Duratio
 use sui_types::{
     base_types::{AuthorityName, ObjectID},
     batch::UpdateItem,
-    crypto::get_key_pair_from_rng,
     messages::ExecutionStatus,
     object::Object,
     utils::{make_committee_key, make_committee_key_num},
@@ -1343,8 +1342,8 @@ pub async fn checkpoint_tests_setup(num_objects: usize, batch_interval: Duration
 
     // Generate a large number of objects for testing
     for _i in 0..num_objects {
-        let (addr1, key1) = get_key_pair_from_rng(&mut rng);
-        let (addr2, _) = get_key_pair_from_rng(&mut rng);
+        let (addr1, key1) = KeyPair::get_key_pair_from_rng(&mut rng);
+        let (addr2, _) = KeyPair::get_key_pair_from_rng(&mut rng);
         let gas_object1 = Object::with_owner_for_testing(addr1);
         let gas_object2 = Object::with_owner_for_testing(addr1);
