+++ crates/sui-core/src/generate_format.rs
@@ -10,11 +10,12 @@ use move_core_types::{
 use pretty_assertions::assert_str_eq;
 use serde_reflection::{Registry, Result, Samples, Tracer, TracerConfig};
 use signature::Signer;
+use sui_types::crypto::KeyPair;
 use std::{fs::File, io::Write};
 use sui_types::{
     base_types::{self, ObjectDigest, ObjectID, TransactionDigest, TransactionEffectsDigest},
     batch::UpdateItem,
-    crypto::{get_key_pair, AuthoritySignature, Signature},
+    crypto::{AuthoritySignature, Signature},
     messages::{
         CallArg, ExecutionFailureStatus, ExecutionStatus, ObjectArg, ObjectInfoRequestKind,
         SingleTransactionKind, TransactionKind,
@@ -31,7 +32,7 @@ fn get_registry() -> Result<Registry> {
     // tracer.trace_value(&mut samples, ...)?;
     // with all the base types contained in messages, especially the ones with custom serializers;
     // or involving generics (see [serde_reflection documentation](https://novifinancial.github.io/serde-reflection/serde_reflection/index.html)).
-    let (addr, kp) = get_key_pair();
+    let (addr, kp) = KeyPair::get_key_pair();
     let pk = kp.public_key_bytes();
     tracer.trace_value(&mut samples, &addr)?;
     tracer.trace_value(&mut samples, &kp)?;
