+++ crates/sui-core/src/unit_tests/move_integration_tests.rs
@@ -10,7 +10,7 @@ use crate::authority::authority_tests::{
 use move_package::BuildConfig;
 use sui_types::{
     crypto::KeyPair,
-    crypto::{get_key_pair, Signature},
+    crypto::{Signature},
     event::{Event, EventType, TransferType},
     messages::ExecutionStatus,
     object::OBJECT_START_VERSION,
@@ -23,7 +23,7 @@ const MAX_GAS: u64 = 10000;
 
 #[tokio::test]
 async fn test_object_wrapping_unwrapping() {
-    let (sender, sender_key) = get_key_pair();
+    let (sender, sender_key) = KeyPair::get_key_pair();
     let gas = ObjectID::random();
     let authority = init_state_with_ids(vec![(sender, gas)]).await;
 
@@ -203,7 +203,7 @@ async fn test_object_wrapping_unwrapping() {
 
 #[tokio::test]
 async fn test_object_owning_another_object() {
-    let (sender, sender_key) = get_key_pair();
+    let (sender, sender_key) = KeyPair::get_key_pair();
     let gas = ObjectID::random();
     let authority = init_state_with_ids(vec![(sender, gas)]).await;
