+++ crates/sui-types/src/crypto_traits.rs
@@ -1,0 +1,43 @@
+use anyhow::Error;
+use serde::Serialize;
+use sha3::Sha3_256;
+
+/// Something that we know how to hash and sign.
+pub trait Signable<W> {
+    fn write(&self, writer: &mut W);
+}
+pub trait SignableBytes
+where
+    Self: Sized,
+{
+    fn from_signable_bytes(bytes: &[u8]) -> Result<Self, anyhow::Error>;
+}
+/// Activate the blanket implementation of `Signable` based on serde and BCS.
+/// * We use `serde_name` to extract a seed from the name of structs and enums.
+/// * We use `BCS` to generate canonical bytes suitable for hashing and signing.
+pub trait BcsSignable: Serialize + serde::de::DeserializeOwned {}
+
+impl<T, W> Signable<W> for T
+where
+    T: BcsSignable,
+    W: std::io::Write,
+{
+    fn write(&self, writer: &mut W) {
+        let name = serde_name::trace_name::<Self>().expect("Self must be a struct or an enum");
+        // Note: This assumes that names never contain the separator `::`.
+        write!(writer, "{}::", name).expect("Hasher should not fail");
+        bcs::serialize_into(writer, &self).expect("Message serialization should not fail");
+    }
+}
+
+impl<T> SignableBytes for T
+where
+    T: BcsSignable,
+{
+    fn from_signable_bytes(bytes: &[u8]) -> Result<Self, Error> {
+        // Remove name tag before deserialization using BCS
+        let name = serde_name::trace_name::<Self>().expect("Self must be a struct or an enum");
+        let name_byte_len = format!("{}::", name).bytes().len();
+        Ok(bcs::from_bytes(&bytes[name_byte_len..])?)
+    }
+}
