+++ crates/sui-types/src/unit_tests/base_types_tests.rs
@@ -9,9 +9,9 @@ use std::str::FromStr;
 use base64ct::{Base64, Encoding};
 use move_binary_format::file_format;
 
-use crate::crypto::{get_key_pair_from_bytes, AuthoritySignature, KeyPair};
+use crate::crypto::{AuthoritySignature, KeyPair};
 use crate::{
-    crypto::{get_key_pair, BcsSignable, Signature},
+    crypto::{BcsSignable, Signature},
     gas_coin::GasCoin,
     object::Object,
     SUI_FRAMEWORK_ADDRESS,
@@ -31,8 +31,8 @@ impl BcsSignable for Bar {}
 
 #[test]
 fn test_signatures() {
-    let (addr1, sec1) = get_key_pair();
-    let (addr2, _sec2) = get_key_pair();
+    let (addr1, sec1) = KeyPair::get_key_pair();
+    let (addr2, _sec2) = KeyPair::get_key_pair();
 
     let foo = Foo("hello".into());
     let foox = Foo("hellox".into());
@@ -288,7 +288,7 @@ fn test_transaction_digest_serde_human_readable() {
 
 #[test]
 fn test_signature_serde_not_human_readable() {
-    let (_, key) = get_key_pair();
+    let (_, key) = KeyPair::get_key_pair();
     let sig = AuthoritySignature::new(&Foo("some data".to_string()), &key);
     let serialized = bincode::serialize(&sig).unwrap();
     let bcs_serialized = bcs::to_bytes(&sig).unwrap();
@@ -301,7 +301,7 @@ fn test_signature_serde_not_human_readable() {
 
 #[test]
 fn test_signature_serde_human_readable() {
-    let (_, key) = get_key_pair();
+    let (_, key) = KeyPair::get_key_pair();
     let sig = AuthoritySignature::new(&Foo("some data".to_string()), &key);
     let serialized = serde_json::to_string(&sig).unwrap();
     assert_eq!(
@@ -350,7 +350,7 @@ const SAMPLE_ADDRESS: &str = "ee0437cf625b77af4d12bff98af1a88332b00638";
 
 // Derive a sample address and public key tuple from KeyPair bytes.
 fn derive_sample_address() -> (SuiAddress, KeyPair) {
-    let (address, pub_key) = get_key_pair_from_bytes(&[
+    let (address, pub_key) = KeyPair::get_key_pair_from_bytes(&[
         10, 112, 5, 142, 174, 127, 187, 146, 251, 68, 22, 191, 128, 68, 84, 13, 102, 71, 77, 57,
         92, 154, 128, 240, 158, 45, 13, 123, 57, 21, 194, 214, 189, 215, 127, 86, 129, 189, 1, 4,
         90, 106, 17, 10, 123, 200, 40, 18, 34, 173, 240, 91, 213, 72, 183, 249, 213, 210, 39, 181,
