+++ crates/sui/src/bin/bench_configure.rs
@@ -5,8 +5,9 @@ use clap::*;
 use std::path::Path;
 use sui_config::genesis_config::{AccountConfig, GenesisConfig, ObjectConfigRange};
 use sui_config::Config;
+use sui_types::crypto::KeyPair;
 
-use sui_types::{base_types::ObjectID, crypto::get_key_pair};
+use sui_types::{base_types::ObjectID};
 
 #[derive(Debug, Parser)]
 #[clap(
@@ -35,7 +36,7 @@ fn main() {
     // For each load gen, create an account
     for _ in 0..bch.number_of_generators {
         // Create a keypair for this account
-        let (account_address, account_keypair) = get_key_pair();
+        let (account_address, account_keypair) = KeyPair::get_key_pair();
 
         // Populate the range configs
         let range_cfg = ObjectConfigRange {
